Add pyramid create
