+++ ndpyramid/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 
+from .create import pyramid_create
 from .coarsen import pyramid_coarsen
 from .reproject import pyramid_reproject
 from .regrid import pyramid_regrid
